+++ iam_pb.tf
@@ -168,16 +168,13 @@ data "aws_iam_policy_document" "permissions-boundary-policy" {
       "iam:CreateServiceLinkedRole",
     ]
     resources = [
-      "arn:aws:iam::${local.account_id}:role/aws-service-role/eks-nodegroup.amazonaws.com/AWSServiceRoleForAmazonEKSNodegroup",
-      "arn:aws:iam::${local.account_id}:role/aws-service-role/elasticloadbalancing.amazonaws.com/AWSServiceRoleForElasticLoadBalancing",
+      for k, v in local.service_linked_role_definitions :
+      "arn:aws:iam::${local.account_id}:role/aws-service-role/${v.service_name}/${v.role_name}"
     ]
     condition {
       test     = "StringEqualsIfExists"
       variable = "iam:AWSServiceName"
-      values = [
-        "eks-nodegroup.amazonaws.com",
-        "elasticloadbalancing.amazonaws.com",
-      ]
+      values   = [for k, v in local.service_linked_role_definitions : v.service_name]
     }
   }